(feat) Rename components with .js to .jsx extension
