+++ client/components/NavBar.jsx
