+++ client/containers/RedditContainer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import Reddit from '../components/Reddit';
+import Reddit from '../components/Reddit.jsx';
 
 function mapStateToProps({ reddit }) {
   return { reddit };
