+++ client/containers/RootContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import * as actions from '../actions/index.js';
-import RootComponent from '../components/Root.js';
+import RootComponent from '../components/Root.jsx';
 
 const mapDispatchToProps = (dispatch) => {
   return {
