+++ client/routes.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, IndexRedirect } from 'react-router';
-import CoreLayout from './components/CoreLayout';
+import CoreLayout from './components/CoreLayout.jsx';
 
 export default (
   <Route path="/" component={CoreLayout}>
